Remove styleResizable.
